+++ src/Common/Internal/ServiceRestProxy.php
@@ -503,10 +503,12 @@ class ServiceRestProxy extends RestProxy
      *
      * @return string
      */
-    public function groupQueryValues($values)
+    public static function groupQueryValues($values)
     {
         Validate::isArray($values, 'values');
         $joined = Resources::EMPTY_STRING;
+
+        sort($values);
 
         foreach ($values as $value) {
             if (!is_null($value) && !empty($value)) {